requirements.txt
# Core ML & Data
pandas
numpy
scikit-learn
xgboost

# Live GFS Data
herbie-data
xarray
requests

# Verification (from your script)
feedparser
beautifulsoup4
transformers
torch

streamlit

folium
streamlit-folium
reverse_geocoder